Recipe names were not formatted correctly
When adding multiple recipes to index, the text was not formatted into a new line.
Instead it was displayed in 1 line.

To solve this problem I utilized the <p></p> (paragraph) element. This alone did not
solve the problem so I learned about the <br> tag to add a text break. This resolved it.

Summary:
     Figured out how to wrap other elements with text elements/tags.
     Learned the <br> tag.